FiltreringIAutomatiskBehandling -> FiltreringsreglerUtil, endret noen variabelnavn i filtreringsregler
