+++ src/test/kotlin/no/nav/familie/ba/sak/kjerne/fødselshendelse/FødselshendelseIntegrasjonTest.kt
@@ -18,7 +18,7 @@ import no.nav.familie.ba.sak.integrasjoner.pdl.internal.DødsfallData
 import no.nav.familie.ba.sak.integrasjoner.pdl.internal.IdentInformasjon
 import no.nav.familie.ba.sak.integrasjoner.pdl.internal.PersonInfo
 import no.nav.familie.ba.sak.integrasjoner.pdl.internal.VergeData
-import no.nav.familie.ba.sak.kjerne.automatiskvurdering.EvaluerFiltreringsreglerService
+import no.nav.familie.ba.sak.kjerne.automatiskvurdering.FiltreringsreglerService
 import no.nav.familie.ba.sak.kjerne.behandling.NyBehandlingHendelse
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingRepository
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingResultat
@@ -94,7 +94,7 @@ class FødselshendelseIntegrasjonTest(
         private val evaluerFiltreringsreglerForFødselshendelse: EvaluerFiltreringsreglerForFødselshendelse,
 
         @Autowired
-        private val evaluerFiltreringsreglerService: EvaluerFiltreringsreglerService,
+        private val filtreringsreglerService: FiltreringsreglerService,
 
         @Autowired
         private val vedtakService: VedtakService,
@@ -130,7 +130,7 @@ class FødselshendelseIntegrasjonTest(
             stegService,
             vedtakService,
             evaluerFiltreringsreglerForFødselshendelse,
-            evaluerFiltreringsreglerService,
+            filtreringsreglerService,
             taskRepository,
             personopplysningerService,
             vilkårsvurderingRepository,
